requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
PyPDF2
faiss-cpu
python-dotenv
flask